+++ notification-server/resources/schema.sql
@@ -1,6 +1,7 @@
 -- accounts created.
 CREATE TABLE IF NOT EXISTS account_device_mapping (
   id SERIAL8 PRIMARY KEY,
-  address BYTEA NOT NULL UNIQUE,
-  device_id VARCHAR NOT NULL UNIQUE
+  address BYTEA NOT NULL,
+  device_id VARCHAR NOT NULL,
+  UNIQUE (address, device_id)
 );
