+++ notification-server/src/bin/service.rs
@@ -1,11 +1,5 @@
 use clap::Parser;
 use concordium_rust_sdk::{
-    cis2,
-    cis2::Event,
-    types::{
-        AccountTransactionEffects, BlockItemSummaryDetails::AccountTransaction,
-        ContractTraceElement,
-    },
     v2::{Client, Endpoint},
 };
 use dotenv::dotenv;
@@ -13,6 +7,7 @@ use tonic::{
     codegen::{http, tokio_stream::StreamExt},
     transport::ClientTlsConfig,
 };
+use notification_server::processor::process;
 
 #[derive(Debug, Parser)]
 struct Args {
@@ -31,40 +26,6 @@ struct Args {
         env = "DB_CONNECTION"
     )]
     db_connection: tokio_postgres::config::Config,
-}
-
-fn get_cis2_events_addresses(effects: &AccountTransactionEffects) -> Option<Vec<String>> {
-    match &effects {
-        AccountTransactionEffects::ContractUpdateIssued { effects } => Some(
-            effects
-                .iter()
-                .flat_map(|effect| match effect {
-                    ContractTraceElement::Updated { data } => data
-                        .events
-                        .iter()
-                        .map(|event| match cis2::Event::try_from(event) {
-                            Ok(Event::Transfer { to, .. }) => Some(to.to_string()),
-                            Ok(Event::Mint { amount, .. }) => Some(amount.to_string()),
-                            _ => None,
-                        })
-                        .filter(Option::is_some)
-                        .collect(),
-                    _ => None,
-                })
-                .collect(),
-        ),
-        _ => None,
-    }
-}
-
-fn is_notification_emitting_transaction_effect(effects: &AccountTransactionEffects) -> bool {
-    matches!(
-        effects,
-        AccountTransactionEffects::AccountTransfer { .. }
-            | AccountTransactionEffects::AccountTransferWithMemo { .. }
-            | AccountTransactionEffects::TransferredWithSchedule { .. }
-            | AccountTransactionEffects::TransferredWithScheduleAndMemo { .. }
-    )
 }
 
 #[tokio::main(flavor = "multi_thread")]
@@ -96,28 +57,7 @@ async fn main() -> anyhow::Result<()> {
             .get_block_transaction_events(block_hash)
             .await?
             .response;
-        let addresses: Vec<String> = transactions
-            .filter_map(Result::ok)
-            .filter_map(|t| match t.details {
-                AccountTransaction(ref account_transaction) => {
-                    if is_notification_emitting_transaction_effect(&account_transaction.effects) {
-                        Some(
-                            t.affected_addresses()
-                                .into_iter()
-                                .map(|addr| addr.to_string())
-                                .collect::<Vec<_>>(),
-                        )
-                    } else {
-                        get_cis2_events_addresses(&account_transaction.effects)
-                    }
-                }
-                _ => None,
-            })
-            .collect::<Vec<Vec<String>>>()
-            .await
-            .into_iter()
-            .flatten()
-            .collect();
+        let addresses = process(transactions).await;
         println!("Addresses: {:#?}", addresses);
     }
     Ok(())
